Add "Age" and "Language" to harassment list
I think it makes sense to add "age" and "language" to these lists. Language discrimination refers to the unfair treatment of an individual based solely upon the characteristics of their speech; such as, accent, size of vocabulary, and syntax. It can also involve a person's ability or inability to use one language instead of another.