Update harassment policy (#10)
* Update harassment list

A couple of weeks ago we updated the list of protected groups under the harassment definition, but we forgot to update the list in the opening statement. This PR maintains consistency between those two lists.

* Add "Age" and "Language" to harassment list

I think it makes sense to add "age" and "language" to these lists. Language discrimination refers to the unfair treatment of an individual based solely upon the characteristics of their speech; such as, accent, size of vocabulary, and syntax. It can also involve a person's ability or inability to use one language instead of another.